pyproject.toml
[tool.poetry]
name = "main-street-app"
version = "0.1.0"
description = ""
authors = ["slaav <slaava.novitskiy175@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.29.0"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
uvicorn = {extras = ["standart"], version = "^0.30.1"}
alembic = "^1.13.2"
ruff = "^0.5.1"
mypy = "^1.10.1"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
httpx = "^0.27.0"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
fastapi-redis-cache = "^0.2.5"
pre-commit = "^3.7.1"



[tool.ruff]
exclude = [".venv", ".env", "migrations"]
extend-ignore = ["E712"]
# Set the maximum line length to 79.
line-length = 79

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
extend-select = ["E501"]

[tool.pytest.ini_options]

minversion = "6.0"
testpaths = [
    "tests",
    "integration",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
